Move operation to its own file
